requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.6
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
thefuzz>=0.20.0
python-Levenshtein>=0.23.0
groq>=0.9.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
pytest
flake8
